refactor: move base dto
