+++ src/modules/common/base.dto.ts
