+++ src/modules/user/dto/update-user-password.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString } from 'class-validator';
-import { BaseDTO } from 'src/base-dto';
+import { BaseDTO } from 'src/modules/common/base.dto';
 
 export class UpdateUserPasswordDTO extends BaseDTO {
   @IsString()
